Más de la persistencia. El XML ya casi tiene la forma definitiva y el cargado (salvo por cosas que faltan definir: colecciones, Point, etc) está funcionando correctamente.

+++ tp2-codigo-fuente/tp2/modelo/Vuelo.java
@@ -88,7 +88,7 @@ public abstract class Vuelo implements IGuardable{
 		this.objetoVolador = (ObjetoVolador) ReconstructorDesdeXml.generarObjeto(atributos.get("objetoVolador"));
 		this.iniciado = (Boolean) ReconstructorDesdeXml.generarObjeto(atributos.get("iniciado"));
 		this.origen = (Point) ReconstructorDesdeXml.generarObjeto(atributos.get("origen"));
-		this.trayectoriaDeVuelo = (Integer) ReconstructorDesdeXml.generarObjeto(atributos.get("trayectoriaDeVuelo"));
+		this.trayectoriaDeVuelo = (Double) ReconstructorDesdeXml.generarObjeto(atributos.get("trayectoriaDeVuelo"));
 		return this;
 	}
 }
